@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, EventEmitter, Input, Output, signal, SimpleChanges } from '@angular/core';
+import { Component, signal, SimpleChanges, input, output } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { AllSvgComponent } from "../svg/all-svg/all-svg.component";
 
@@ -10,22 +10,23 @@ import { AllSvgComponent } from "../svg/all-svg/all-svg.component";
   styleUrl: './selector.component.css'
 })
 export class SelectorComponent {
-  @Input() placeholder: string = 'Select an option';
-  @Input() options: any[] = [];
-  @Input() selectedId: any = null; // Preselected ID from the parent
-  @Output() selectionChange = new EventEmitter<any>();
+  readonly placeholder = input<string>('Select an option');
+  readonly options = input<any[]>([]);
+  readonly selectedId = input<any>(null); // Preselected ID from the parent
+  readonly selectionChange = output<any>();
 
   inputValue: string = '';
   selected = signal<any>('');
   open: boolean = false;
 
   ngOnChanges(changes: SimpleChanges): void {
-    if (this.selectedId === "") {
+    const selectedId = this.selectedId();
+    if (selectedId === "") {
       this.selected.set(null);
     }
 
-    if (changes['selectedId'] && this.selectedId !== null) {
-      this.setSelectedById(this.selectedId);
+    if (changes['selectedId'] && selectedId !== null) {
+      this.setSelectedById(selectedId);
     }
   }
 
@@ -34,7 +35,7 @@ export class SelectorComponent {
   }
 
   setSelectedById(id: number): void {
-    const selectedOption = this.options.find(option => option.id == id);
+    const selectedOption = this.options().find(option => option.id == id);
     if (selectedOption) {
       this.selected.set(selectedOption);
       this.selectionChange.emit(this.selected());
