@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, ElementRef, inject, QueryList, signal, ViewChild, ViewChildren } from '@angular/core';
+import { Component, ElementRef, inject, signal, viewChildren, viewChild } from '@angular/core';
 import { ToastSuccessComponent } from '../../../../shared/toast/toast-success/toast-success.component';
 import { FieldComponent } from '../../../../shared/field/field.component';
 import { SearchComponent } from '../../../../shared/svg/search/search.component';
@@ -26,8 +26,8 @@ export class VendorEntryComponent {
   private searchQuery$ = new BehaviorSubject<string>('');
   isLoading$: Observable<any> | undefined;
   hasError$: Observable<any> | undefined;
-  @ViewChildren('inputRef') inputRefs!: QueryList<ElementRef>;
-  @ViewChild('searchInput') searchInput!: ElementRef<HTMLInputElement>;
+  readonly inputRefs = viewChildren<ElementRef>('inputRef');
+  readonly searchInput = viewChild.required<ElementRef<HTMLInputElement>>('searchInput');
   isSubmitted = false;
 
   form = this.fb.group({
@@ -42,7 +42,7 @@ export class VendorEntryComponent {
 
     // Focus on the search input when the component is initialized
     setTimeout(() => {
-      const inputs = this.inputRefs.toArray();
+      const inputs = this.inputRefs();
       inputs[0].nativeElement.focus();
     }, 10); // Delay to ensure the DOM is updated
   }
@@ -83,7 +83,7 @@ export class VendorEntryComponent {
   handleEnterKey(event: Event, currentIndex: number) {
     const keyboardEvent = event as KeyboardEvent;
     event.preventDefault();
-    const allInputs = this.inputRefs.toArray();
+    const allInputs = this.inputRefs();
     const inputs = allInputs.filter((i: any) => !i.nativeElement.disabled);
 
     if (currentIndex + 1 < inputs.length) {
@@ -100,7 +100,7 @@ export class VendorEntryComponent {
 
     if (event.key === 'Tab') {
       event.preventDefault();
-      const inputs = this.inputRefs.toArray();
+      const inputs = this.inputRefs();
       inputs[0].nativeElement.focus();
     } else if (event.key === 'ArrowDown') {
       event.preventDefault(); // Prevent default scrolling behavior
@@ -182,7 +182,7 @@ export class VendorEntryComponent {
 
     // Focus the 'Name' input field after patching the value
     setTimeout(() => {
-      const inputs = this.inputRefs.toArray();
+      const inputs = this.inputRefs();
       inputs[0].nativeElement.focus();
     }, 0); // Delay to ensure the DOM is updated
   }
