@@ -1,4 +1,4 @@
-import { Component, ElementRef, inject, QueryList, signal, ViewChild, ViewChildren } from '@angular/core';
+import { Component, ElementRef, inject, signal, viewChildren, viewChild } from '@angular/core';
 import { FormControl, NonNullableFormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { UserAccessService } from '../../../services/user-access.service';
 import { UserAccessTreeComponent } from "../../shared/user/user-access-tree/user-access-tree.component";
@@ -30,8 +30,8 @@ export class UsersComponent {
   userAccessTree = signal<any[]>([]);
   isLoading$: Observable<any> | undefined;
   hasError$: Observable<any> | undefined;
-  @ViewChildren('inputRef') inputRefs!: QueryList<ElementRef>;
-  @ViewChild('searchInput') searchInput!: ElementRef<HTMLInputElement>;
+  readonly inputRefs = viewChildren<ElementRef>('inputRef');
+  readonly searchInput = viewChild.required<ElementRef<HTMLInputElement>>('searchInput');
   isSubmitted = false;
 
   form = this.fb.group({
@@ -48,7 +48,7 @@ export class UsersComponent {
 
     // Focus on the search input when the component is initialized
     setTimeout(() => {
-      const inputs = this.inputRefs.toArray();
+      const inputs = this.inputRefs();
       inputs[0].nativeElement.focus();
     }, 10);
   }
@@ -86,7 +86,7 @@ export class UsersComponent {
   handleEnterKey(event: Event, currentIndex: number) {
     const keyboardEvent = event as KeyboardEvent;
     event.preventDefault();
-    const allInputs = this.inputRefs.toArray();
+    const allInputs = this.inputRefs();
     const inputs = allInputs.filter((i: any) => !i.nativeElement.disabled);
 
     if (currentIndex + 1 < inputs.length) {
@@ -103,7 +103,7 @@ export class UsersComponent {
 
     if (event.key === 'Tab') {
       event.preventDefault();
-      const inputs = this.inputRefs.toArray();
+      const inputs = this.inputRefs();
       inputs[0].nativeElement.focus();
     } else if (event.key === 'ArrowDown') {
       event.preventDefault(); // Prevent default scrolling behavior
@@ -183,7 +183,7 @@ export class UsersComponent {
 
     // Focus the 'username' input field after patching the value
     setTimeout(() => {
-      const inputs = this.inputRefs.toArray();
+      const inputs = this.inputRefs();
       inputs[0].nativeElement.focus();
     }, 0); // Delay to ensure the DOM is updated
   }
